Refactor the game instance (backend) into its own file
